Done customer API, transaction API, contact form

+++ spring-boot-web/src/main/java/com/javaweb/repository/CustomerRepository.java
@@ -5,5 +5,5 @@ import com.javaweb.repository.custom.CustomerRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<CustomerEntity, Long>, CustomerRepositoryCustom {
-    CustomerEntity findByIdAndIsActive(Long id, Long isActive);
+    CustomerEntity findByIdAndIsActive(Long id, Integer isActive);
 }
